Update superagent-proxy to address security advisory
Security advisory: https://www.npmjs.com/advisories/1184

The security advisory was already addressed in commit 8e592df, but
something about the lockfile seems to confuse the audit API (or we're
running into a bug with resolutions). This resolution is simpler, and
accomplishes the same goal of updating the vulnerable package.