Add account types (#167)
* Add account types

Introduces a concept of "account types" to the AccountsController.

When requesting the permission to `wallet_manageIdentities` (now known
as `wallet_manageAccounts_*`, where `*` is an account type identifier
string. Currently this string is `bip44:60` for
Ethereum, with the intention of rendering any bip44 protocol's name
given the second segment's identifier.

Also introduces `wallet_accounts_*`, where `*` is again an account type
identifier. This method is meant to so that `eth_accounts` can be an
alias for `wallet_accounts_bip44:60`, except that it also supports
paramters.

If any parameters are present in a call to `wallet_accounts_*`, then
those parameters will be passed to the handler for the account defined
by the `from` value on its `AccountMessageHandler` function, allowing
account type snaps to define their own interfaces, and the accounts
permission to allow passing through messages related to the permitted
accounts.

Will benefit from a number of improvements:
- [ ] Rendering protocol names from slip44 identifier.
- [ ] Allow selecting accounts of non-eth protocols.
- [ ] Redirect `eth_accounts` at `wallet_accounts_*`.
- [ ] See if we can redirect all classic eth signing methods through
this new general purpose send method (allows keyrings to iterate faster
than the keyring-controller).

* Linted

* Fix slip44

* Linted

* Linted

* Linted

* Fix bugs, make service registration use protocol name