Minor grammar change
Some of these workings sounds like an AI or Belgian person did them. I added less awkward Dutch phrasings.

The word `gekopieerde` is an adverb, in the UI it is used as a verb and should thus be `gekopieerd`.

The word `vrij` means free, but in `line 3, line 6` context you strictly need `blootleggen, roughly `to lay bare`. The use of `weergeven` on `line 9` however is more is less unnatural (blootleggen sounds like "would you like to lay bare your accounts to this website").

`Aangepaste` doesn't mean custom but rather 'edited'. I used 'handmatig' which means 'manual'. It is the closest word in Dutch I can think of.

`one-stop-bestemming` is not a thing. I used 'all in one store'.

## Questions

Is ` "Hier is een lijst !!!!"` intentional?